+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 
 - Deps: update `@angular` to 17.2.3.
 - Deps: update `@ionic` to 7.7.3.
+- Deps: update `@types/node` to 20.11.22.